# contracts/openapi/professor.yaml
openapi: 3.1.0
info:
  title: eduanima-professor API
  version: 1.0.0
  description: |
    Professor API (HTTP/JSON + SSE)

    ## 通信スタック
    - Frontend ↔ Professor: HTTP/JSON + SSE（永続）
    - Professor ↔ Librarian: gRPC双方向ストリーミング（Phase 1から）

    ## Phase 1: ローカル開発
    - 全リクエストで user_id='00000000-0000-0000-0000-000000000001' を固定使用
    - `X-Dev-User: dev-user` ヘッダーによる固定ユーザー識別
    - Phase 2でSSO認証実装時に /v1/auth/dev-login を削除し、Authorization: Bearer <token> に移行

    ## エンドポイント設計原則
    - リソースは名詞・複数形（subjects / materials / chats）
    - 階層ネスト: subjects 配下に materials・chats を配置
    - SSE: POST /v1/subjects/{id}/chats → Content-Type: text/event-stream で直接ストリーム
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/subjects/{subject_id}/chats:
    post:
      tags: [Chats]
      summary: 質問送信（SSE ストリーミング）
      description: |
        質問を送信し、Server-Sent Events でリアルタイム応答を受信。
        内部では Professor が Librarian に gRPC 双方向ストリーミングで推論を委譲する。
        Phase 1: X-Dev-User ヘッダーによる固定ユーザー
        Phase 2以降: Bearer トークンによるSSO認証

        SSEイベント種別:
          thinking  → AI Agentが戦略立案中（プログレスバー表示用）
          searching → 資料を検索中（クエリ文字列を含む）
          evidence  → 根拠資料を選定完了（ファイル名・ページ・抜粋）
          chunk     → 回答テキスト断片（Markdown形式で逐次配信）
          done      → 完了（chat_idを含む）
          error     → エラー発生（再試行可能）
      parameters:
        - $ref: '#/components/parameters/SubjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question:
                  type: string
                  example: "決定係数の計算式を教えてください"
                parent_chat_id:
                  type: string
                  format: uuid
                  nullable: true
                  description: ヒアリング判断で選択肢から選んだ場合、親チャットIDを指定
      responses:
        '200':
          description: SSEストリーミング開始
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  SSEフォーマット（各行は "data: {json}" 形式）
              examples:
                thinking:
                  summary: 戦略立案中
                  value: |
                    data: {"type":"thinking","content":"検索戦略を立案中..."}
                searching:
                  summary: 検索中
                  value: |
                    data: {"type":"searching","query":"決定係数 計算式"}
                evidence:
                  summary: 根拠資料選定
                  value: |
                    data: {"type":"evidence","material_id":"uuid","name":"lecture03.pdf","page":12,"excerpt":"決定係数 R² は..."}
                chunk:
                  summary: 回答テキスト断片
                  value: |
                    data: {"type":"chunk","content":"決定係数 R² の計算式は..."}
                done:
                  summary: 完了
                  value: |
                    data: {"type":"done","chat_id":"uuid"}
                error:
                  summary: エラー
                  value: |
                    data: {"type":"error","code":"SEARCH_TIMEOUT","request_id":"uuid"}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      tags: [Chats]
      summary: 会話履歴一覧取得
      description: |
        指定した科目の会話履歴一覧を取得。
        Web版の「会話履歴」表示に使用。
      parameters:
        - $ref: '#/components/parameters/SubjectId'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: cursor
          in: query
          required: false
          description: カーソルベースページネーション
          schema:
            type: string
      responses:
        '200':
          description: 会話履歴リスト
          content:
            application/json:
              schema:
                type: object
                required: [chats]
                properties:
                  chats:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatSummary'
                  next_cursor:
                    type: string
                    nullable: true
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/subjects/{subject_id}/chats/{chat_id}:
    get:
      tags: [Chats]
      summary: 会話詳細取得
      parameters:
        - $ref: '#/components/parameters/SubjectId'
        - $ref: '#/components/parameters/ChatId'
      responses:
        '200':
          description: 会話詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/subjects/{subject_id}/chat/completions:
    post:
      tags: [OpenAI-Compat]
      summary: OpenAI Chat Completions 互換（RAG 質問応答）
      description: |
        OpenAI Chat Completions API 互換エンドポイント。
        OpenAI SDK を使って eduanimaR professor に接続するには base_url を
        "http://host/api/v1/subjects/{subject_id}" に設定する。

        messages 配列の最後の role="user" メッセージが質問として使用される。

        stream=true（デフォルト）の場合:
          SSE で OpenAI delta 形式にて回答をストリーミング。
          標準の delta.content に回答テキストを、
          eduanima_event 拡張フィールドに thinking/searching/evidence 中間イベントを付与。

        stream=false の場合:
          全チャンクをバッファして chat.completion JSON オブジェクトを返す。
          eduanima_sources 拡張フィールドに引用ソース情報を付与。

        Python SDK 例:
          from openai import OpenAI
          client = OpenAI(
              base_url="http://localhost:8080/api/v1/subjects/{subject_id}",
              api_key="dev-token"
          )
          response = client.chat.completions.create(
              model="professor",
              messages=[{"role": "user", "content": "光合成とは?"}],
              stream=True
          )
      parameters:
        - $ref: '#/components/parameters/SubjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenAIChatRequest'
      responses:
        '200':
          description: |
            stream=true: SSE ストリーミング（OpenAI delta 形式）
            stream=false: chat.completion JSON オブジェクト
          content:
            text/event-stream:
              schema:
                type: string
                description: OpenAI delta chunk SSE（stream=true 時）
              examples:
                role_chunk:
                  summary: 最初のチャンク（role アナウンス）
                  value: |
                    data: {"id":"chatcmpl-abc123","object":"chat.completion.chunk","created":1234567890,"model":"professor","choices":[{"index":0,"delta":{"role":"assistant","content":""},"finish_reason":null}]}
                thinking_chunk:
                  summary: 中間イベント（thinking）
                  value: |
                    data: {"id":"chatcmpl-abc123","object":"chat.completion.chunk","created":1234567890,"model":"professor","choices":[{"index":0,"delta":{},"finish_reason":null}],"eduanima_event":{"type":"thinking","data":{"message":"Analyzing your question..."}}}
                content_chunk:
                  summary: 回答テキストチャンク
                  value: |
                    data: {"id":"chatcmpl-abc123","object":"chat.completion.chunk","created":1234567890,"model":"professor","choices":[{"index":0,"delta":{"content":"光合成とは"},"finish_reason":null}]}
                done_chunk:
                  summary: 完了チャンク + センチネル
                  value: |
                    data: {"id":"chatcmpl-abc123","object":"chat.completion.chunk","created":1234567890,"model":"professor","choices":[{"index":0,"delta":{},"finish_reason":"stop"}]}

                    data: [DONE]
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAIChatCompletion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    SubjectId:
      name: subject_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ChatId:
      name: chat_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ChatSummary:
      type: object
      required: [chat_id, subject_id, question, created_at]
      properties:
        chat_id:
          type: string
          format: uuid
        subject_id:
          type: string
          format: uuid
        question:
          type: string
          example: "決定係数の計算式を教えてください"
        answered_at:
          type: string
          format: date-time
          nullable: true
          description: null の場合は未完了（タイムアウト/エラー）
        created_at:
          type: string
          format: date-time
    ChatDetail:
      type: object
      required: [chat_id, subject_id, question, created_at]
      properties:
        chat_id:
          type: string
          format: uuid
        subject_id:
          type: string
          format: uuid
        question:
          type: string
        answer:
          type: string
          nullable: true
          description: 最終回答テキスト（SSE完了後に保存）
        sources:
          type: array
          nullable: true
          description: 参照した資料の一覧
          items:
            type: object
            required: [material_id, name]
            properties:
              material_id:
                type: string
                format: uuid
              name:
                type: string
              page:
                type: integer
                nullable: true
              excerpt:
                type: string
        feedback:
          type: string
          enum: [good, bad]
          nullable: true
        parent_chat_id:
          type: string
          format: uuid
          nullable: true
          description: ヒアリング判断による選択肢選択の親チャット
        answered_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time

    # ─── OpenAI 互換スキーマ（ADR 005）─────────────────────────
    OpenAIChatRequest:
      type: object
      required: [messages]
      properties:
        model:
          type: string
          description: |
            professor では無視される（subject_id は URL パスから取得）。
            任意の文字列を指定可能（例: "professor", "gpt-4o"）。
          example: "professor"
        messages:
          type: array
          description: 会話履歴。最後の role="user" メッセージが質問として使用される。
          items:
            type: object
            required: [role, content]
            properties:
              role:
                type: string
                enum: [user, assistant, system]
              content:
                type: string
        stream:
          type: boolean
          nullable: true
          description: |
            true（デフォルト）: SSE ストリーミング応答
            false: バッファして chat.completion JSON を返す
          default: true
    OpenAIChatCompletion:
      type: object
      description: OpenAI chat.completion オブジェクト（stream=false 時）
      required: [id, object, created, model, choices, usage]
      properties:
        id:
          type: string
          example: "chatcmpl-abc12345"
        object:
          type: string
          enum: [chat.completion]
        created:
          type: integer
          format: int64
          description: Unix タイムスタンプ
        model:
          type: string
          example: "professor"
        choices:
          type: array
          items:
            type: object
            required: [index, message, finish_reason]
            properties:
              index:
                type: integer
              message:
                type: object
                required: [role, content]
                properties:
                  role:
                    type: string
                    enum: [assistant]
                  content:
                    type: string
                    description: 完全な回答テキスト
              finish_reason:
                type: string
                enum: [stop]
        usage:
          type: object
          description: トークン使用量（professor では常に 0）
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer
        eduanima_sources:
          type: array
          description: |
            eduanimaR 固有の引用ソース情報（拡張フィールド）。
            標準 OpenAI クライアントはこのフィールドを無視する。
          items:
            type: object
            properties:
              file_name:
                type: string
              excerpt:
                type: string
              why_relevant:
                type: string
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              description: アプリケーションエラーコード（docs/03_integration/ERROR_CODES.md 参照）
              example: "VALIDATION_FAILED"
            message:
              type: string
              description: ユーザー向けメッセージ（内部詳細・Secretsを含まない）
              example: "Validation failed"
            details:
              type: object
              additionalProperties: true
            request_id:
              type: string
              format: uuid